chore: improve Spinner exports
